Add ToString representation to Node
Provides a default representation to Node using ToString trait.
The to_string method is a shortcut, for serializing the Node using
default serializiation of html5ever nodes.

With ToString trait implemented one can replace:

    let document = Html::from_string(html).parse(&arena);
    let mut serialized = Vec::new();
    serialize(&mut serialized, Default::default()).unwrap();
    assert_eq!(String::from_utf8(serialized).unwrap(), r"<!DOCTYPE ..");

with:

    let document = Html::from_string(html).parse(&arena);
    assert_eq!(document.to_string(), r"<!DOCTYPE ..");

If for some reason parsing or serializing encounters an error it will
return empty String.